+++ app/src/main/java/com/practicum/playlist5/media/domain/api/FavouriteRepository.kt
@@ -12,5 +12,5 @@ interface FavouriteRepository {
 
     fun getID(trackId:Int): Flow<Track>
 
-    suspend fun isFavourite(trackId: Int):Boolean
+    suspend fun isFavourite(trackId: Long):Boolean
 }
